chore(json-ld): update and move JsonLdUtil
